\name{ASRR-package}
\alias{ASRR-package}
\alias{ASRR}
\docType{package}
\title{
Applied Sociological Research with R
}
\description{
Miscellaneous functions for sociological research.}
\details{
\tabular{ll}{
Package: \tab ASRR\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2009-01-10\cr
License: \tab BSD License\cr
LazyLoad: \tab yes\cr
URL: \tab \url{http://asrr.r-forge.r-project.org/} \cr
}
}
\author{
  Ronggui HUANG

  Maintainer: Ronggui HUANG <ronggui.huang@gmail.com>
}

\keyword{ package }

